refactor: split ui model into focused modules
