+++ internal/ui/mode_diff.go
@@ -1,0 +1,36 @@
+package ui
+
+import tea "github.com/charmbracelet/bubbletea"
+
+// Diff mode key handling and viewport delegation.
+
+func (m Model) updateDiffMode(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
+	switch msg.String() {
+	case "q", "ctrl+c":
+		return m, tea.Quit
+	case "esc", "h", "left":
+		m.mode = modeFileList
+		return m, nil
+	case "n":
+		return m.nextFile()
+	case "p":
+		return m.prevFile()
+	case "e":
+		if m.cursor < len(m.files) {
+			m.SelectedFile = m.files[m.cursor].change.Path
+		}
+		return m, tea.Quit
+	case "b":
+		return m.enterBranchMode()
+	case "tab":
+		return m.toggleStage()
+	case "v":
+		m.splitDiff = !m.splitDiff
+		m.prevCurs = -1
+		m.lastDiffContent = ""
+		return m, tea.Batch(m.loadDiffCmd(true), m.saveSplitPrefCmd())
+	}
+	var cmd tea.Cmd
+	m.viewport, cmd = m.viewport.Update(msg)
+	return m, cmd
+}
